Clean up deploy process and messaging
