+++ cli-scripts/deploy-prod.js
@@ -1,9 +1,10 @@
+/* eslint-disable no-console */
 const colors = require("colors/safe");
 const prompt = require("prompt");
 const promptRun = require("prompt-run");
 const simpleGit = require("simple-git");
 const path = require("path");
-const getLatestReleaseTag = require("./git-tag-proc");
+const { getLatestCommitsSinceLastRelease } = require("./deploy-utils");
 
 const git = simpleGit(path.resolve(__dirname, ".."));
 
@@ -20,41 +21,6 @@ const checkIfOnMaster = async () => {
   return true;
 };
 
-// Fetches all the commits since the last release tag if there are no commits
-// since the last release it returns false otherwise prints out all the commits
-const getLatestCommitsSinceLastRelease = async () => {
-  try {
-    const latestRelease = getLatestReleaseTag();
-    const latestCommits = await git.log(latestRelease, "HEAD");
-    if (latestCommits.all.length === 0) {
-      console.log(colors.red("No Changes found since last release\n"));
-      return false;
-    }
-    console.log(colors.red("Changes since last release\n"));
-    prettyPrintCommitLogs(latestCommits.all);
-    return true;
-  } catch (e) {
-    console.log(
-      colors.bgRed(
-        `Error occured when trying to parse commits.\nThere is likely something wrong with your commit history:\n ${e}`
-      )
-    );
-    return false;
-  }
-};
-
-// Prints out all the commits since the commit body is optional it will conditionally print it
-const prettyPrintCommitLogs = (latestCommits) => {
-  latestCommits.forEach((commit) => {
-    console.log(`hash: ${colors.yellow(commit.hash)}`);
-    console.log(`author: ${colors.cyan(commit.author_name)}`);
-    console.log(`email: ${colors.cyan(commit.author_email)}`);
-    console.log(`change: ${colors.cyan(commit.message)}`);
-    if (commit.body) console.log(`body: ${colors.cyan(commit.body)}`);
-    console.log("\n");
-  });
-};
-
 // Runs the script to build and deploy to production
 const deployProcess = () => {
   promptRun({
@@ -68,14 +34,22 @@ const deployProcess = () => {
     childProcess.on("close", () => {
       console.log("Deploy Complete! 🎉");
       console.log("Don't forget to push tags");
-      console.log(colors.green("git push upstream && git push upstream --tags"));
+      console.log(
+        colors.green("git push upstream && git push upstream --tags")
+      );
     });
   });
 };
 
 const deployProd = async () => {
   const onMaster = await checkIfOnMaster();
-  const anyChangesFound = await getLatestCommitsSinceLastRelease();
+  const latestCommits = await getLatestCommitsSinceLastRelease();
+  const anyChangesFound = latestCommits.length > 0;
+  if (anyChangesFound) {
+    console.log(latestCommits);
+  } else {
+    console.log(colors.red("No Changes found"));
+  }
   if (onMaster && anyChangesFound) {
     prompt.start();
     prompt.get(
